requirements.txt
torch
torchvision
torchaudio
scikit-learn
numpy
scipy
tqdm
matplotlib
pillow
